resolved conflicts + fixed some minor bugs because of last refeactoring

+++ Projeto/src/main/java/voluntier/util/produces/CauseDataReturn.java
@@ -9,7 +9,7 @@ public class CauseDataReturn extends PicturesReturn {
 	public String id;
 	public String name;
 	public Integer goal;
-	public Integer raised;
+	public Double raised;
 	public Integer num_donations;
 	public String website;
 	public String description;
@@ -24,7 +24,7 @@ public class CauseDataReturn extends PicturesReturn {
 		this.id = cause.getString(DB_Cause.ID);
 		this.name = cause.getString(DB_Cause.NAME);
 		this.goal = (int) cause.getLong(DB_Cause.GOAL);
-		this.raised = (int) cause.getDouble(DB_Cause.RAISED);
+		this.raised = cause.getDouble(DB_Cause.RAISED);
 		this.description = cause.getString(DB_Cause.DESCRIPTION);
 		this.description = cause.getString(DB_Cause.WEBSITE);
 		this.company_name = cause.getString(DB_Cause.WITH);
